Move trie code to its own file
